requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

aiohappyeyeballs==2.6.1 ; python_version >= "3.10"
aiohttp==3.12.15 ; python_version >= "3.10"
aiosignal==1.4.0 ; python_version >= "3.10"
aiosmtplib==4.0.2 ; python_version >= "3.10"
aiosqlite==0.19.0 ; python_version >= "3.10"
alembic==1.16.5 ; python_version >= "3.10"
annotated-types==0.7.0 ; python_version >= "3.10"
anyio==4.11.0 ; python_version >= "3.10"
argon2-cffi-bindings==25.1.0 ; python_version >= "3.10"
argon2-cffi==25.1.0 ; python_version >= "3.10"
async-timeout==5.0.1 ; python_version >= "3.10" and python_version < "3.12.0"
asyncpg==0.29.0 ; python_version >= "3.10"
attrs==25.3.0 ; python_version >= "3.10"
bcrypt==4.3.0 ; python_version >= "3.10"
cachetools==6.2.0 ; python_version >= "3.10"
certifi==2025.8.3 ; python_version >= "3.10"
cffi==2.0.0 ; python_version >= "3.10"
charset-normalizer==3.4.3 ; python_version >= "3.10"
click==8.3.0 ; python_version >= "3.10"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version >= "3.10"
cryptography==46.0.2 ; python_version >= "3.10"
dnspython==2.8.0 ; python_version >= "3.10"
ecdsa==0.19.1 ; python_version >= "3.10"
email-validator==2.3.0 ; python_version >= "3.10"
exceptiongroup==1.3.0 ; python_version == "3.10"
fastapi==0.118.0 ; python_version >= "3.10"
filelock==3.19.1 ; python_version >= "3.10"
frozenlist==1.7.0 ; python_version >= "3.10"
fsspec==2025.9.0 ; python_version >= "3.10"
google-ai-generativelanguage==0.6.15 ; python_version >= "3.10"
google-api-core==2.25.1 ; python_version >= "3.10"
google-api-python-client==2.183.0 ; python_version >= "3.10"
google-auth-httplib2==0.2.0 ; python_version >= "3.10"
google-auth==2.41.1 ; python_version >= "3.10"
google-generativeai==0.8.5 ; python_version >= "3.10"
googleapis-common-protos==1.70.0 ; python_version >= "3.10"
greenlet==3.2.4 ; python_version >= "3.10"
grpcio-status==1.71.2 ; python_version >= "3.10"
grpcio==1.75.1 ; python_version >= "3.10"
h11==0.16.0 ; python_version >= "3.10"
hf-xet==1.1.10 ; python_version >= "3.10" and (platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "arm64" or platform_machine == "aarch64")
httplib2==0.31.0 ; python_version >= "3.10"
httptools==0.6.4 ; python_version >= "3.10"
huggingface-hub==0.35.3 ; python_version >= "3.10"
idna==3.10 ; python_version >= "3.10"
jinja2==3.1.6 ; python_version >= "3.10"
joblib==1.5.2 ; python_version >= "3.10"
mako==1.3.10 ; python_version >= "3.10"
markupsafe==3.0.3 ; python_version >= "3.10"
minio==7.2.18 ; python_version >= "3.10"
mpmath==1.3.0 ; python_version >= "3.10"
multidict==6.6.4 ; python_version >= "3.10"
networkx==3.4.2 ; python_version == "3.10"
networkx==3.5 ; python_version >= "3.11"
numpy==1.26.4 ; python_version >= "3.10"
packaging==25.0 ; python_version >= "3.10"
passlib==1.7.4 ; python_version >= "3.10"
pgvector==0.2.5 ; python_version >= "3.10"
pillow==11.3.0 ; python_version >= "3.10"
propcache==0.3.2 ; python_version >= "3.10"
proto-plus==1.26.1 ; python_version >= "3.10"
protobuf==5.29.5 ; python_version >= "3.10"
psycopg2-binary==2.9.10 ; python_version >= "3.10"
pyasn1-modules==0.4.2 ; python_version >= "3.10"
pyasn1==0.6.1 ; python_version >= "3.10"
pycparser==2.23 ; python_version >= "3.10" and implementation_name != "PyPy"
pycryptodome==3.23.0 ; python_version >= "3.10"
pydantic-core==2.33.2 ; python_version >= "3.10"
pydantic-settings==2.11.0 ; python_version >= "3.10"
pydantic==2.11.9 ; python_version >= "3.10"
pyparsing==3.2.5 ; python_version >= "3.10"
python-dotenv==1.1.1 ; python_version >= "3.10"
python-jose==3.5.0 ; python_version >= "3.10"
python-multipart==0.0.20 ; python_version >= "3.10"
pyyaml==6.0.3 ; python_version >= "3.10"
regex==2025.9.18 ; python_version >= "3.10"
requests==2.32.5 ; python_version >= "3.10"
rsa==4.2 ; python_version >= "3.14"
rsa==4.9.1 ; python_version >= "3.10" and python_version <= "3.13"
safetensors==0.6.2 ; python_version >= "3.10"
scikit-learn==1.7.2 ; python_version >= "3.10"
scipy==1.15.3 ; python_version == "3.10"
scipy==1.16.2 ; python_version >= "3.11"
sentence-transformers==5.1.1 ; python_version >= "3.10"
six==1.17.0 ; python_version >= "3.10"
sniffio==1.3.1 ; python_version >= "3.10"
sqlalchemy==2.0.43 ; python_version >= "3.10"
starlette==0.48.0 ; python_version >= "3.10"
sympy==1.14.0 ; python_version >= "3.10"
threadpoolctl==3.6.0 ; python_version >= "3.10"
tokenizers==0.22.1 ; python_version >= "3.10"
tomli==2.2.1 ; python_version == "3.10"
torch==2.2.2+cpu ; python_version >= "3.10"
tqdm==4.67.1 ; python_version >= "3.10"
transformers==4.56.2 ; python_version >= "3.10"
typing-extensions==4.15.0 ; python_version >= "3.10"
typing-inspection==0.4.2 ; python_version >= "3.10"
uritemplate==4.2.0 ; python_version >= "3.10"
urllib3==2.5.0 ; python_version >= "3.10"
uvicorn==0.37.0 ; python_version >= "3.10"
uvloop==0.21.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10"
watchfiles==1.1.0 ; python_version >= "3.10"
websockets==15.0.1 ; python_version >= "3.10"
yarl==1.20.1 ; python_version >= "3.10"